Refactor find_client_info
