+++ test/common.hpp
@@ -11,6 +11,7 @@
 
 #include <chrono>
 #include <memory>
+#include <string_view>
 
 // The timeout for tests involving communication to a real server.
 // Some tests use a longer timeout by multiplying this value by some
@@ -35,3 +36,7 @@ void run(
    boost::redis::config cfg = make_test_config(),
    boost::system::error_code ec = boost::asio::error::operation_aborted,
    boost::redis::operation op = boost::redis::operation::receive);
+
+// Finds a value in the output of the CLIENT INFO command
+// format: key1=value1 key2=value2
+std::string_view find_client_info(std::string_view client_info, std::string_view key);
