+++ test/test_conn_setup.cpp
@@ -29,21 +29,6 @@ using boost::system::error_code;
 
 namespace {
 
-// Finds a value in the output of the CLIENT INFO command
-// format: key1=value1 key2=value2
-std::string_view find_client_info(std::string_view client_info, std::string_view key)
-{
-   std::string prefix{key};
-   prefix += '=';
-
-   auto const pos = client_info.find(prefix);
-   if (pos == std::string_view::npos)
-      return {};
-   auto const pos_begin = pos + prefix.size();
-   auto const pos_end = client_info.find(' ', pos_begin);
-   return client_info.substr(pos_begin, pos_end - pos_begin);
-}
-
 // Creates a user with a known password. Harmless if the user already exists
 void setup_password()
 {
